Add cast() method to fix test

+++ lib/src/resources/news_api_provider.dart
@@ -19,7 +19,7 @@ class NewsApiProvider {
 
     final ids = json.decode(response.body);
 
-    return ids;
+    return ids.cast<int>();
   }
 
   Future<ItemModel> fetchItem(int id) async {
